small optimization of memory usage of the example game and some inlining

+++ app/src/main/java/com/jetpackcompose/playground/compose_game_bench/data/RaycastScreenColumnInfo.kt
@@ -3,7 +3,7 @@ package com.jetpackcompose.playground.compose_game_bench.data
 data class RaycastScreenColumnInfo(
     var wallHeight: Double = 0.0,
     var colorIntensity: Int = 0,
-    var xOffset: Float = 0f,
+    var xOffset: Int = 0,
     var worldTextureOffset: Double = 0.0,
     var hitWallNumber: Int = 0
 )
